[libcdtr] -Fix: debug prints should use double and %f for printing coords to be portable across different coord types
git-svn-id: svn://repo.hu/pcb-rnd/trunk@33799 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_3rd/libcdtr/cdt.c
@@ -327,7 +327,7 @@ static void check_adjacent_triangle(triangle_t *adj_t, edge_t *adj_e, point_t *n
 void dump_pointlist(pointlist_node_t *list)
 {
 	POINTLIST_FOREACH(p, list)
-		fprintf(stderr, "\t{pos=(%d, %d) adj_t=(", p->pos.x, p->pos.y);
+		fprintf(stderr, "\t{pos=(%f, %f) adj_t=(", (double)p->pos.x, (double)p->pos.y);
 		TRIANGLELIST_FOREACH(t, p->adj_triangles)
 			fprintf(stderr, "%p", (void*)t);
 			if(_node_->next != NULL)
@@ -346,8 +346,8 @@ void dump_pointlist(pointlist_node_t *list)
 void dump_edgelist(edgelist_node_t *list)
 {
 	EDGELIST_FOREACH(e, list)
-		fprintf(stderr, "\t{endp1=(%p (pos=(%d, %d))) endp2=(%p (pos=(%d, %d))) adj_t=(",
-						(void*)e->endp[0], e->endp[0]->pos.x, e->endp[0]->pos.y, (void*)e->endp[1], e->endp[1]->pos.x, e->endp[1]->pos.y);
+		fprintf(stderr, "\t{endp1=(%p (pos=(%f, %f))) endp2=(%p (pos=(%f, %f))) adj_t=(",
+						(void*)e->endp[0], (double)e->endp[0]->pos.x, (double)e->endp[0]->pos.y, (void*)e->endp[1], (double)e->endp[1]->pos.x, (double)e->endp[1]->pos.y);
 		fprintf(stderr, "%p, %p)", (void*)e->adj_t[0], (void*)e->adj_t[1]);
 		fprintf(stderr, " %s}\n", e->is_constrained ? "constrained" : "");
 	EDGELIST_FOREACH_END();
@@ -356,8 +356,8 @@ void dump_edgelist(edgelist_node_t *list)
 void dump_trianglelist(trianglelist_node_t *list)
 {
 	TRIANGLELIST_FOREACH(t, list)
-		fprintf(stderr, "\t{p1=(%d, %d) p2=(%d, %d) p3=(%d, %d)",
-						t->p[0]->pos.x, t->p[0]->pos.y, t->p[1]->pos.x, t->p[1]->pos.y, t->p[2]->pos.x, t->p[2]->pos.y);
+		fprintf(stderr, "\t{p1=(%f, %f) p2=(%f, %f) p3=(%f, %f)",
+						(double)t->p[0]->pos.x, (double)t->p[0]->pos.y, (double)t->p[1]->pos.x, (double)t->p[1]->pos.y, (double)t->p[2]->pos.x, (double)t->p[2]->pos.y);
 		fprintf(stderr, " adj_e=(%p, %p, %p)", (void*)t->e[0], (void*)t->e[1], (void*)t->e[2]);
 		fprintf(stderr, " adj_t=(%p, %p, %p)}\n", (void*)t->adj_t[0], (void*)t->adj_t[1], (void*)t->adj_t[2]);
 	TRIANGLELIST_FOREACH_END();
